refactor component_metadata: move implementation to core
This follows mavlink_ftp_client, so that other plugins can access component
metadata as well.
The Plugin API remains unchanged.
