+++ src/mavsdk/core/file_cache.h
